+++ src/main/resources/application.properties
@@ -45,7 +45,7 @@ s3.secret-key=${S3_SECRET_KEY}
 
 # Set to false during development to skip authentication
 # MUST be true before any demo or production deployment!
-app.security.enabled=true
+app.security.enabled=false
 
 # Production default ? always secure
 cookie.secure=true
